Finish serialization

+++ Ruby/Hangman/lib/game.rb
@@ -73,7 +73,7 @@ class Game
     Time.now.to_i - @start_time
   end
 
-  def to_json(*options)
+  def to_json(*_options)
     current_state = {
       current_guess: @current_guess,
       guessed_letters: @guessed_letters,
@@ -86,19 +86,81 @@ class Game
     JSON.pretty_generate(current_state)
   end
 
+  def save_exists?
+    File.exist?('save.json')
+  end
+
   def save
+    continue_serialization = 'y'
+    continue_game = ''
+
+    if save_exists?
+      print 'Warning - this will overwrite your existing save. '
+
+      loop do
+        print 'Continue (y/n)? '
+        continue_serialization = gets.chomp.downcase
+        puts
+
+        unless %w[y n].include?(continue_serialization)
+          puts 'Invalid - please input y or n'
+          next
+        end
+
+        break
+      end
+    end
+
+    return unless continue_serialization == 'y'
+
     File.open('save.json', 'w') do |file|
       file.write(to_json)
     end
+
+    print 'File saved. '
+
+    loop do
+      print 'End game (y/n)? '
+      continue_game = gets.chomp.downcase
+      puts
+
+      unless %w[y n].include?(continue_game)
+        puts 'Invalid - please input y or n'
+        next
+      end
+
+      break
+    end
+
+    exit if continue_game == 'y'
   end
 
   def load_save
     saved_state = {}
+    continue_deserialization = false
+
     File.open('save.json', 'r') do |file|
       saved_state = JSON.parse(file.read)
     end
 
     # Deserialization
+    print 'This will end your current game. '
+
+    loop do
+      print 'Continue (y/n)? '
+      continue_deserialization = gets.chomp.downcase
+      puts
+
+      unless %w[y n].include?(continue_deserialization)
+        puts 'Invalid - please input y or n'
+        next
+      end
+
+      break
+    end
+
+    return unless continue_deserialization == 'y'
+
     @current_guess = saved_state['current_guess']
     @guessed_letters = saved_state['guessed_letters']
     @number_of_guesses = saved_state['number_of_guesses']
@@ -116,7 +178,7 @@ class Game
 
     loop do # Primary game loop
       # puts "You are on guess number #{@number_of_guesses} (#{@word})"
-      puts "You are on guess number #{@number_of_guesses} - #{elapsed_seconds} seconds elapsed"
+      puts "You are on guess number #{@number_of_guesses} - Seconds elapsed: #{elapsed_seconds} (#{@word})"
 
       unless @guessed_letters.empty?
         print 'Letters guessed: '
@@ -128,7 +190,12 @@ class Game
         puts
       end
 
-      print 'Enter your guess, or press ENTER to guess the word: '
+      if save_exists?
+        print 'Enter your guess, press ENTER to guess the word, enter 0 to load save, or enter 1 to save: '
+      else
+        print 'Enter your guess, press ENTER to guess the word, or enter 1 to save: '
+      end
+
       @current_guess = gets.chomp.downcase
       puts
 
@@ -140,6 +207,12 @@ class Game
         else
           next
         end
+      elsif @current_guess == '0'
+        load_save
+        next
+      elsif @current_guess == '1'
+        save
+        next
       elsif @current_guess.length != 1 || !@current_guess.ord.between?(97, 122)
         puts 'Invalid guess!'
         next
@@ -158,8 +231,6 @@ class Game
         puts "Took #{@number_of_guesses} guesses and #{elapsed_seconds} seconds"
         break
       end
-
-      save
     end
   end
 end
